Implantado todas as funçoes para adicionar, incluir e excluir e editar.

+++ src/br/com/letscode/cookbook/domain/ModoPreparo.java
@@ -12,7 +12,7 @@ public class ModoPreparo {
         this.descriçãoPasso = descriçãoPasso;
     }
 
-    public int getPasso() {
+    public double getPasso() {
         return passo;
     }
